requirements.txt
pymongo
pymongo[srv]
dnspython
pandas
numpy
ensure
pytest
mysql-connector-python

-e .